WEEK 3/May 22/requirements.txt
PyMuPDF==1.25.5
chromadb==1.0.10
google-generativeai==0.8.5
autogen-core==0.5.7
autogen-agentchat==0.5.7
